--- a/src/Lux.jl
+++ b/src/Lux.jl
@@ -5,7 +5,7 @@ using CUDA
 using CUDA.CUDNN
 # Neural Network Backend
 using NNlib
-import NNlibCUDA: batchnorm, ∇batchnorm, CUDNNFloat
+import LuxLib  ## In v0.5 we can starting `using`. For v0.4, there will be naming conflicts
 # Julia StdLibs
 using Random, Statistics, LinearAlgebra, SparseArrays
 # Parameter Manipulation