Fixes #10 #98

# Performance Updates

Performance only changes for GroupNorm due to our new KA kernels

## Code

```julia
using BenchmarkTools, Lux, Random, CUDA, Zygote

for sz in ((32, 32, 128, 16), (64, 64, 64, 16), (128, 128, 32, 16)), groups in (4, 8, 16)
    gn = GroupNorm(sz[3], groups)
    x = randn(Float32, sz...) |> gpu
    ps, st = Lux.setup(Random.default_rng(), gn) .|> gpu

    t1 = @belapsed CUDA.@sync Lux.apply($gn, $x, $ps, $st)
    Zygote.gradient((x, ps) -> sum(Lux.apply(gn, x, ps, st)[1]), x, ps)
    t2 = @belapsed CUDA.@sync Zygote.gradient((x, ps) -> sum(Lux.apply($gn, x, ps, $st)[1]),
                                              $x, $ps)

    println("$(sz), $(groups), $(t1), $(t2)")
end
```
## Timings

| Size | N Groups | Time Forward (New) | Time Fwd + Bwd (New) | Time Forward (Old) | Time Fwd + Bwd (Old) |
| - | - | - | - | - | - |
(32, 32, 128, 16) | 4 | 0.001171762 | 0.003418343 | 0.001432062 | 0.006942468
(32, 32, 128, 16) | 8 | 0.001179864 | 0.003470376 | 0.001454761 | 0.006201573
(32, 32, 128, 16) | 16 | 0.001208639 | 0.003496287 | 0.001452037 | 0.006244667
(64, 64, 64, 16) | 4 | 0.002307972 | 0.005459583 | 0.002810693 | 0.011882327
(64, 64, 64, 16) | 8 | 0.002342824 | 0.00553138 | 0.002867196 | 0.011813671
(64, 64, 64, 16) | 16 | 0.002331648 | 0.005505468 | 0.002855742 | 0.012082637
(128, 128, 32, 16) | 4 | 0.00444357 | 0.010073916 | 0.005577281 | 0.024195243
(128, 128, 32, 16) | 8 | 0.004467874 | 0.010130907 | 0.005562752 | 0.024235608
(128, 128, 32, 16) | 16 | 0.004479398 | 0.010194882 | 0.005528737 | 0.024055475